requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
gTTS
python-dotenv
google-generativeai